♻ refactor: 공용 컴포넌트 단수명 변경
